+++ RayTracing/HitRecord.h
@@ -16,20 +16,20 @@ class HitRecord{
 	public:
 		HitRecord();
 		HitRecord(float t1, glm::vec4 p1, glm::vec4 normal1, graphics::Material material1);
-		HitRecord(float t1, glm::vec4 p1, glm::vec4 normal1, graphics::Material material1, int sText1, int tText1, graphics::Texture *texture1);
+		HitRecord(float t1, glm::vec4 p1, glm::vec4 normal1, graphics::Material material1, float sText1, float tText1, graphics::Texture *texture1);
 		~HitRecord();
 		void setT(float t1);
 		void setP(glm::vec4 p1);
 		void setNormal(glm::vec4 normal1);
 		void setMaterial(graphics::Material material1);
-		void setTextCoords(int sText1, int tText1);
+		void setTextCoords(float sText1, float tText1);
 		void setTexture(graphics::Texture *texture1);
 		float getT();
 		glm::vec4 getP();
 		glm::vec4 getNormal();
 		graphics::Material getMaterial();
-		int getSTextCoord();
-		int getTTextCoord();
+		float getSTextCoord();
+		float getTTextCoord();
 		graphics::Texture* getTexture();
 
 	private:
@@ -37,8 +37,8 @@ class HitRecord{
 		glm::vec4 p;
 		glm::vec4 normal;
 		graphics::Material material;
-		int sText;
-		int tText;
+		float sText;
+		float tText;
 		graphics::Texture *texture;
 };
 #endif
